feat: refactor ingestion processor by splitting changeset generation and applying into separate packages
