+++ diode-server/reconciler/differ/differ_virt_test.go
@@ -1,4 +1,4 @@
-package changeset_test
+package differ_test
 
 import (
 	"context"
@@ -12,6 +12,7 @@ import (
 	"github.com/netboxlabs/diode/diode-server/netboxdiodeplugin"
 	"github.com/netboxlabs/diode/diode-server/netboxdiodeplugin/mocks"
 	"github.com/netboxlabs/diode/diode-server/reconciler/changeset"
+	"github.com/netboxlabs/diode/diode-server/reconciler/differ"
 )
 
 func TestVirtualizationPrepare(t *testing.T) {
@@ -24,14 +25,14 @@ func TestVirtualizationPrepare(t *testing.T) {
 	}
 	tests := []struct {
 		name                 string
-		ingestEntity         changeset.IngestEntity
+		ingestEntity         differ.IngestEntity
 		retrieveObjectStates []mockRetrieveObjectState
 		wantChangeSet        changeset.ChangeSet
 		wantErr              bool
 	}{
 		{
 			name: "[P1] ingest virtualization.clustergroup with name only - existing object not found - create",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.clustergroup",
 				Entity: &diodepb.Entity{
@@ -73,7 +74,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P1] ingest virtualization.clustergroup with name only - existing object found - do nothing",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.clustergroup",
 				Entity: &diodepb.Entity{
@@ -107,7 +108,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P1] ingest empty virtualization.clustergroup - error",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.clustergroup",
 				Entity: &diodepb.Entity{
@@ -125,7 +126,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P2] ingest virtualization.clustertype with name only - existing object not found - create",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.clustertype",
 				Entity: &diodepb.Entity{
@@ -167,7 +168,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P2] ingest virtualization.clustertype with name only - existing object found - do nothing",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.clustertype",
 				Entity: &diodepb.Entity{
@@ -201,7 +202,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P2] ingest empty virtualization.clustertype - error",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.clustertype",
 				Entity: &diodepb.Entity{
@@ -227,7 +228,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P3] ingest virtualization.cluster with name only - existing object not found - create",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.cluster",
 				Entity: &diodepb.Entity{
@@ -318,7 +319,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P3] ingest virtualization.cluster with name only - existing object found - do nothing",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.cluster",
 				Entity: &diodepb.Entity{
@@ -391,7 +392,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P3] ingest empty virtualization.cluster - error",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.cluster",
 				Entity: &diodepb.Entity{
@@ -409,7 +410,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P4] ingest virtualization.virtualmachine with name only - existing object not found - create",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.virtualmachine",
 				Entity: &diodepb.Entity{
@@ -485,7 +486,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P4] ingest virtualization.virtualmachine with name and cluster - existing objects not found - create",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.virtualmachine",
 				Entity: &diodepb.Entity{
@@ -632,7 +633,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P4] ingest virtualization.virtualmachine with name and existing cluster - existing vm not found - create",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.virtualmachine",
 				Entity: &diodepb.Entity{
@@ -769,7 +770,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P4] ingest virtualization.virtualmachine with name and cluster - existing vm found - create cluster",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.virtualmachine",
 				Entity: &diodepb.Entity{
@@ -927,7 +928,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P4] ingest virtualization.virtualmachine with name only - existing object found - do nothing",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.virtualmachine",
 				Entity: &diodepb.Entity{
@@ -1074,7 +1075,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P4] ingest empty virtualization.virtualmachine - error",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.virtualmachine",
 				Entity: &diodepb.Entity{
@@ -1092,7 +1093,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P5] ingest virtualization.vminterface with name only - existing object not found - create",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.vminterface",
 				Entity: &diodepb.Entity{
@@ -1197,7 +1198,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P5] ingest virtualization.vminterface with name only - existing object found - do nothing",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.vminterface",
 				Entity: &diodepb.Entity{
@@ -1289,7 +1290,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P5] ingest empty virtualization.vminterface - error",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.vminterface",
 				Entity: &diodepb.Entity{
@@ -1307,7 +1308,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P6] ingest virtualization.virtualdisk with name only - existing object not found - create",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.virtualdisk",
 				Entity: &diodepb.Entity{
@@ -1412,7 +1413,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P6] ingest virtualization.virtualdisk with name only and no existing site - existing object not found - create",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.virtualdisk",
 				Entity: &diodepb.Entity{
@@ -1528,7 +1529,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P6] ingest virtualization.virtualdisk with name only - existing object found - do nothing",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.virtualdisk",
 				Entity: &diodepb.Entity{
@@ -1607,7 +1608,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 		},
 		{
 			name: "[P6] ingest empty virtualization.virtualdisk - error",
-			ingestEntity: changeset.IngestEntity{
+			ingestEntity: differ.IngestEntity{
 				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
 				DataType:  "virtualization.virtualdisk",
 				Entity: &diodepb.Entity{
@@ -1642,7 +1643,7 @@ func TestVirtualizationPrepare(t *testing.T) {
 				}, nil)
 			}
 
-			cs, err := changeset.Prepare(tt.ingestEntity, mockClient)
+			cs, err := differ.Diff(tt.ingestEntity, mockClient)
 			if tt.wantErr {
 				require.Error(t, err)
 				return
