+++ diode-server/reconciler/ingestion_processor_internal_test.go
@@ -107,111 +107,6 @@ func TestWriteIngestionLog(t *testing.T) {
 	}
 }
 
-func TestReconcileEntity(t *testing.T) {
-	tests := []struct {
-		name                   string
-		retrieveObjectStateErr error
-		applyErr               error
-		expectedError          bool
-		expectedCS             *changeset.ChangeSet
-	}{
-		{
-			name:          "successful reconciliation",
-			expectedError: false,
-			expectedCS: &changeset.ChangeSet{
-				ChangeSetID: "5663a77e-9bad-4981-afe9-77d8a9f2b8b5",
-				ChangeSet: []changeset.Change{
-					{
-						ChangeID:      "5663a77e-9bad-4981-afe9-77d8a9f2b8b5",
-						ChangeType:    changeset.ChangeTypeCreate,
-						ObjectType:    "dcim.site",
-						ObjectID:      nil,
-						ObjectVersion: nil,
-						Data: &netbox.DcimSite{
-							Name:   "Site A",
-							Slug:   "site-a",
-							Status: (*netbox.DcimSiteStatus)(strPtr(string(netbox.DcimSiteStatusActive))),
-						},
-					},
-				},
-			},
-		},
-		{
-			name:                   "prepare error",
-			retrieveObjectStateErr: errors.New("prepare error"),
-			expectedError:          true,
-		},
-		{
-			name: "apply error",
-			expectedCS: &changeset.ChangeSet{
-				ChangeSetID: "cs123",
-				ChangeSet:   []changeset.Change{},
-			},
-			applyErr:      errors.New("apply error"),
-			expectedError: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx := context.Background()
-
-			// Mock nbClient
-			mockNbClient := new(mnp.NetBoxAPI)
-			logger := slog.New(slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelDebug, AddSource: false}))
-			// Create IngestionProcessor
-			p := &IngestionProcessor{
-				nbClient: mockNbClient,
-				logger:   logger,
-			}
-
-			// Setup mock for RetrieveObjectState
-			if tt.retrieveObjectStateErr != nil {
-				mockNbClient.On("RetrieveObjectState", ctx, mock.Anything).Return(&netboxdiodeplugin.ObjectState{}, tt.retrieveObjectStateErr)
-			} else {
-
-				mockNbClient.On("RetrieveObjectState", ctx, mock.Anything).Return(&netboxdiodeplugin.ObjectState{ObjectType: "dcim.site",
-					ObjectID:       0,
-					ObjectChangeID: 0,
-					Object: &netbox.DcimSiteDataWrapper{
-						Site: nil,
-					}}, nil)
-			}
-			// Setup mock for ApplyChangeSet
-			if tt.expectedCS != nil {
-				mockNbClient.On("ApplyChangeSet", ctx, mock.Anything).Return(&netboxdiodeplugin.ChangeSetResponse{}, tt.applyErr)
-			}
-
-			// Call reconcileEntity
-			ingestEntity := changeset.IngestEntity{
-				RequestID: "cfa0f129-125c-440d-9e41-e87583cd7d89",
-				DataType:  "dcim.site",
-				Entity: &diodepb.Entity{
-					Entity: &diodepb.Entity_Site{
-						Site: &diodepb.Site{
-							Name: "Site A",
-						},
-					},
-				},
-			}
-
-			cs, err := p.reconcileEntity(ctx, ingestEntity)
-
-			if tt.expectedError {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-				require.Equal(t, tt.expectedCS.ChangeSet[0].ChangeType, cs.ChangeSet[0].ChangeType)
-				require.Equal(t, tt.expectedCS.ChangeSet[0].ObjectType, cs.ChangeSet[0].ObjectType)
-				require.Equal(t, tt.expectedCS.ChangeSet[0].Data, cs.ChangeSet[0].Data)
-			}
-
-			// Assert expectations
-			mockNbClient.AssertExpectations(t)
-		})
-	}
-}
-
 func TestHandleStreamMessage(t *testing.T) {
 	tests := []struct {
 		name              string
